Moved save-return logic from unaccompanied to token_based_resume controller
